+++ js/src/dapps/basiccoin/Deployments/deployments.css
@@ -18,7 +18,7 @@
 @import '../_vars.css';
 
 a.link, a.link:hover, a.link:visited {
-  color: $bgEvents;
+  color: $bgDeployments;
   cursor: pointer;
 }
