+++ js/src/dapps/basiccoin/_vars.css
@@ -18,4 +18,4 @@
 $bgEvents: #808;
 $bgDeploy: #088;
 $bgOverview: #080;
-$bgStatus: #f80;
+$bgDeployments: #f80;
